Fix float64 support in Repr.rot() and add comprehensive tests
- Cast generators buffer to input dtype in Repr.rot() for float64 support
- Rewrite test_geometric.py with 124 tests covering:
  - Irrep/ProductIrrep: generators, orthogonality, Lie algebra commutation
  - Repr/ProductRepr: Wigner D-matrices, rotation composition, batching
  - RepNorm: rotation invariance, scaling, edge cases
  - RadialBasisFunctions: all types, cutoffs, learnable params
  - SphericalHarmonic: equivariance, pairwise, zero handling
  - build_knn_graph: correctness, chunking, edge cases
  - EquivariantLinear/LayerNorm: shapes, gradients, bias
  - EquivariantTransformer: equivariance, translation invariance
  - Edge cases: float64, scalar-only, 180° rotations, single point

+++ ciffy/nn/geometric/representations.py
@@ -524,7 +524,7 @@ class Repr(nn.Module):
         """
         # Weight the generators of so(3) by the axis values
         *b, _ = axis.size()
-        gens = (axis @ self.generators).view(*b, self.dim(), self.dim())
+        gens = (axis @ self.generators.to(axis.dtype)).view(*b, self.dim(), self.dim())
 
         # Multiply by the angle and exponentiate to move to SO(3)
         rot = torch.linalg.matrix_exp(angle[..., None, None] * gens)
